Fix #30: Echo manually and fix backspace
So we're getting rid of automatic echoing since pressing weird keys will
lead to cluttering of the terminal. Instead we do things manually by
accepting characters one by one in raw (or cbreak) mode and pushing them
into a buffer then printing that buffer out altogether. This gives that
text editor feel. Of course now the backspace was possible to be handled
manually and pop characters out of the buffer on a backspace encounter
and positioning the cursor properly such that it mimics as if backspace
is working.